services/admin/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
jinja2==3.1.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
